# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusOps Expo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /students:
    get:
      operationId: getStudents
      tags: [students]
      summary: Get all students (Admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
  /hostel/contacts:
    get:
      operationId: getEmergencyContacts
      tags: [hostels]
      summary: Get emergency contacts for user's hostel
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyContact"
    post:
      operationId: createEmergencyContact
      tags: [hostels]
      summary: Create emergency contact (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyContact"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        rollNumber:
          type: string
      required:
        - name
        - email
        - password
        - rollNumber
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, superadmin]
        rollNumber:
          type: string
        hostelId:
          type: string
        roomNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        rollNumber:
          type: string
        hostelId:
          type: string
        hostelName:
          type: string
        roomNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - rollNumber
    CreateStudentRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        rollNumber:
          type: string
        hostelId:
          type: string
        roomNumber:
          type: string
      required:
        - name
        - email
        - password
        - rollNumber
...
    EmergencyContact:
      type: object
      properties:
        id:
          type: string
        hostelId:
          type: string
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        isAvailable24x7:
          type: boolean
      required:
        - id
        - name
        - role
        - phone
    CreateContactRequest:
      type: object
      properties:
        hostelId:
          type: string
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        isAvailable24x7:
          type: boolean
      required:
        - name
        - role
        - phone
...